removed old filers
