+++ src/main/java/com/miiz/controllers/RegisterController.java
@@ -1,6 +1,6 @@
 package com.miiz.controllers;
 
-import com.miiz.AppNew;
+import com.miiz.App;
 import com.miiz.auth.User;
 import com.miiz.auth.UserAuth;
 import com.miiz.database.Database;
@@ -20,9 +20,9 @@ import java.util.Objects;
 
 public class RegisterController {
 
-    private final AppNew app;
+    private final App app;
     private final UserAuth userAuth;
-    public RegisterController(AppNew app, UserAuth userAuth) {
+    public RegisterController(App app, UserAuth userAuth) {
         this.app = app;
         this.userAuth = userAuth;
