--- src/main/java/com/miiz/controllers/appController.java
@@ -1,7 +1,0 @@
-package com.miiz.controllers;
-
-
-
-public class appController {
-
-}
